Separate Models table to hooks, types
