+++ src/components/templates/tables/MonthDailySalaryWithTimeShift.tsx
@@ -24,7 +24,7 @@ const MonthDailySalaryWithTimeShiftTable: React.FC = () => {
       {
         accessorKey: 'operatorName',
         header: 'Оператор',
-        setCellClassName: (): string => 'font-bold',
+        setCellClassName: (): string => 'font-medium',
         asTh: true,
       },
     ];
